requirements.txt
Django==5.2.6
numpy==2.3.3
Pillow==11.3.0
tensorflow==2.20.0
